feature: login page, env, test edited (#65)
