+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\smd99\\today\\slowth\\__tests__\\mocks\\handler\\user-handler.ts":"1","C:\\Users\\smd99\\today\\slowth\\src\\app\\(auth)\\sign-up\\page.tsx":"2","C:\\Users\\smd99\\today\\slowth\\src\\components\\atom\\button\\index.tsx":"3","C:\\Users\\smd99\\today\\slowth\\src\\components\\atom\\input\\index.tsx":"4"},{"size":2287,"mtime":1748323960147,"results":"5","hashOfConfig":"6"},{"size":4054,"mtime":1748324214121,"results":"7","hashOfConfig":"6"},{"size":1778,"mtime":1748321130454,"results":"8","hashOfConfig":"6"},{"size":1314,"mtime":1748322591283,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","suppressedMessages":"12","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1mz997y",{"filePath":"13","messages":"14","suppressedMessages":"15","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","suppressedMessages":"18","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","suppressedMessages":"21","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\smd99\\today\\slowth\\__tests__\\mocks\\handler\\user-handler.ts",[],[],"C:\\Users\\smd99\\today\\slowth\\src\\app\\(auth)\\sign-up\\page.tsx",[],[],"C:\\Users\\smd99\\today\\slowth\\src\\components\\atom\\button\\index.tsx",[],[],"C:\\Users\\smd99\\today\\slowth\\src\\components\\atom\\input\\index.tsx",[],[]]
+[{"C:\\Users\\smd99\\today\\slowth\\__tests__\\mocks\\handler\\user-handler.ts":"1","C:\\Users\\smd99\\today\\slowth\\src\\app\\(auth)\\sign-up\\page.tsx":"2","C:\\Users\\smd99\\today\\slowth\\src\\components\\atom\\button\\index.tsx":"3","C:\\Users\\smd99\\today\\slowth\\src\\components\\atom\\input\\index.tsx":"4","C:\\Users\\smd99\\today\\slowth\\src\\effect\\auth\\fetch-user\\index.ts":"5","C:\\Users\\smd99\\today\\slowth\\src\\effect\\auth\\login\\index.ts":"6","C:\\Users\\smd99\\today\\slowth\\src\\effect\\auth\\sign-up\\index.ts":"7","C:\\Users\\smd99\\today\\slowth\\src\\types\\env.d.ts":"8","C:\\Users\\smd99\\today\\slowth\\vitest.config.ts":"9","C:\\Users\\smd99\\today\\slowth\\vitest.setup.ts":"10"},{"size":2287,"mtime":1748323960147,"results":"11","hashOfConfig":"12"},{"size":3808,"mtime":1748334333362,"results":"13","hashOfConfig":"12"},{"size":1778,"mtime":1748321130454,"results":"14","hashOfConfig":"12"},{"size":1314,"mtime":1748322591283,"results":"15","hashOfConfig":"12"},{"size":251,"mtime":1748333225687,"results":"16","hashOfConfig":"12"},{"size":481,"mtime":1748333236205,"results":"17","hashOfConfig":"12"},{"size":484,"mtime":1748333242670,"results":"18","hashOfConfig":"12"},{"size":175,"mtime":1748334309198,"results":"19","hashOfConfig":"12"},{"size":592,"mtime":1748334162787,"results":"20","hashOfConfig":"12"},{"size":296,"mtime":1748333317106,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","suppressedMessages":"24","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1mz997y",{"filePath":"25","messages":"26","suppressedMessages":"27","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","suppressedMessages":"30","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","suppressedMessages":"33","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","suppressedMessages":"36","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"37","messages":"38","suppressedMessages":"39","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","suppressedMessages":"42","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"43","messages":"44","suppressedMessages":"45","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"46","messages":"47","suppressedMessages":"48","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"49","messages":"50","suppressedMessages":"51","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\smd99\\today\\slowth\\__tests__\\mocks\\handler\\user-handler.ts",[],[],"C:\\Users\\smd99\\today\\slowth\\src\\app\\(auth)\\sign-up\\page.tsx",[],[],"C:\\Users\\smd99\\today\\slowth\\src\\components\\atom\\button\\index.tsx",[],[],"C:\\Users\\smd99\\today\\slowth\\src\\components\\atom\\input\\index.tsx",[],[],"C:\\Users\\smd99\\today\\slowth\\src\\effect\\auth\\fetch-user\\index.ts",[],[],"C:\\Users\\smd99\\today\\slowth\\src\\effect\\auth\\login\\index.ts",[],[],"C:\\Users\\smd99\\today\\slowth\\src\\effect\\auth\\sign-up\\index.ts",[],[],"C:\\Users\\smd99\\today\\slowth\\src\\types\\env.d.ts",[],[],"C:\\Users\\smd99\\today\\slowth\\vitest.config.ts",[],[],"C:\\Users\\smd99\\today\\slowth\\vitest.setup.ts",[],[]]
