+++ src/app/(auth)/login/page.tsx
@@ -1,0 +1,5 @@
+import { LoginForm } from "@/components/organisms/loginForm";
+
+export default function LoginPage() {
+  return <LoginForm />;
+}
