+++ src/effect/auth/fetch-user/index.ts
@@ -1,8 +1,7 @@
-import { TEAM_ID } from "@/constant/test";
-
 export const fetchUser = async () => {
   const response = await fetch(
-    process.env.NEXT_PUBLIC_API_URL + `/${TEAM_ID}/auths/user`,
+    process.env.NEXT_PUBLIC_API_URL +
+      `/${process.env.NEXT_PUBLIC_TEAM_ID}/auths/user`,
   );
   const data = await response.json();
