+++ src/effect/auth/login/index.spec.ts
@@ -5,6 +5,6 @@ import { TEST_LOGIN_USER } from "@/constant/test";
 describe("login", () => {
   test("should return a user", async () => {
     const response = await login(TEST_LOGIN_USER);
-    expect(response).toEqual({ message: "로그인 성공" });
+    expect(response).toHaveProperty("token");
   });
 });
