+++ src/effect/auth/sign-up/index.ts
@@ -1,9 +1,9 @@
-import { TEAM_ID } from "@/constant/test";
 import { SignUpFormInput } from "@/entity/user";
 
 export const signUp = async (data: SignUpFormInput) => {
   const response = await fetch(
-    process.env.NEXT_PUBLIC_API_URL + `/${TEAM_ID}/auths/signup`,
+    process.env.NEXT_PUBLIC_API_URL +
+      `/${process.env.NEXT_PUBLIC_TEAM_ID}/auths/signup`,
     {
       method: "POST",
       mode: "cors",
